% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDegToKM.R
\name{convertDegToKM}
\alias{convertDegToKM}
\title{Convert a set of coordinates in degrees into a new set of coordinates in kilometers.}
\usage{
convertDegToKM(
  loc,
  crs =
    sp::CRS("+units=km +proj=utm +zone=37 +ellps=clrk80 +towgs84=-160,-6,-302,0,0,0,0 +no_defs")
)
}
\arguments{
\item{loc}{A two column matrix of coordinates (The first column is longitude and the second column is latitude).}
}
\value{
A two column matrix of coordinates (The first column is easting and the second column is northing).
}
\description{
Convert a set of coordinates in degrees into a new set of coordinates in kilometers.
}
\examples{
convertDegToKM(loc = NULL)
}
