% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDegToMollweide.R
\name{convertDegToMollweide}
\alias{convertDegToMollweide}
\title{Convert a set of coordinates in degrees into a new set of coordinates in Mollweide coordinate system.}
\usage{
convertDegToMollweide(loc)
}
\arguments{
\item{loc}{A two column matrix of coordinates (The first column is longitude and the second column is latitude).}
}
\value{
A two column matrix of coordinates in Mollweide coordinate system.
}
\description{
Convert a set of coordinates in degrees into a new set of coordinates in Mollweide coordinate system.
}
\examples{
convertDegToMollweide(loc = NULL)
}
