% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertKMToDeg.R
\name{convertKMToDeg}
\alias{convertKMToDeg}
\title{Convert a set of coordinates in kilometers into a new set of coordinates in degrees.}
\usage{
convertKMToDeg(loc)
}
\arguments{
\item{loc}{A two column matrix of coordinates (the first column is easting and the second column is northing).}
}
\value{
A two column matrix of coordinates (the first column is longitude and the second column is latitude).
}
\description{
Convert a set of coordinates in kilometers into a new set of coordinates in degrees.
}
\examples{
convertKMToDeg(loc = NULL)
}
