% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatern.R
\name{covMatern}
\alias{covMatern}
\title{Create a Matern covariance matrix.}
\usage{
covMatern(dMat, range, stdDev)
}
\arguments{
\item{dMat}{Distance matrix.}

\item{range}{Spatial range in kilometers.}

\item{stdDev}{Marginal variance.}
}
\value{
Matern covariance matrix.
}
\description{
Create a Matern covariance matrix.
}
\examples{
covMatern(dMat = NULL, range = NULL, stdDev = NULL)
}
