% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridCountry.R
\name{gridCountry}
\alias{gridCountry}
\title{Create a set of prediction locations, chosen randomly from a prediction grid}
\usage{
gridCountry(admin0, m, n, nPred)
}
\arguments{
\item{admin0}{A SpatialPolygonsdataframe object reresenting the country (admin0) borders}

\item{m}{A value representing the length of the segments that the longitude range of the country will be divided into}

\item{n}{A value representing the length of the segments that the latitude range of the country will be divided into}
}
\value{
A data frame containing the coordinates (in degrees and in kilometers) of a set of prediction points
}
\description{
Create a set of prediction locations, chosen randomly from a prediction grid
}
\examples{
gridCountry(admin0 = NULL, m = NULL, n = NULL, nPred = NULL)
}
