% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshCountry.R
\name{meshCountry}
\alias{meshCountry}
\title{Creates a constrained refined Delaunay triangulation mesh based on the country borders.}
\usage{
meshCountry(admin0, max.edge, offset)
}
\arguments{
\item{admin0}{A SpatialPolygonsDataFrame representing the country borders.}

\item{max.edge}{max.edge values for the inla.mesh.2d object}

\item{offset}{offset value for the inla.mesh.2d object}
}
\value{
A constrained refined Delaunay triangulation mesh created based on the country borders.
}
\description{
Creates a constrained refined Delaunay triangulation mesh based on the country borders.
}
\examples{
mesh.s <- meshCountry(admin0 = admin0, max.edge = c(25, 50), offset = -.08)
}
