% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMesh.R
\name{plotMesh}
\alias{plotMesh}
\title{Create a mesh based on the country borders.}
\usage{
plotMesh(admin0, mesh, name, dir)
}
\arguments{
\item{admin0}{A SpatialPolygonsDataFrame representing the country borders.}

\item{mesh}{The triangulation mesh}

\item{name}{The preferred name for the plot}

\item{dir}{The preferred directory that the plot will be saved into}
}
\value{
Saves a pdf file showing the country (admin0) borders plotted onto the mesh, to the preferred directory
}
\description{
Create a mesh based on the country borders.
}
\examples{
plotMesh(admin0 = NULL, mesh = NULL, name = NULL, dir = NULL)
}
