% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.res.R
\name{print.res}
\alias{print.res}
\title{Prints the output of estimateModel() function.}
\usage{
print.res(x, ...)
}
\arguments{
\item{x}{A list containing the model estimation output, returned by estimateModel function.}

\item{...}{other}
}
\value{
Prints the model parameter estimates and the corresponding credible interval lengths.
}
\description{
Prints the output of estimateModel() function.
}
\examples{
\dontrun{
print.res(est = est)
}
}
