% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDegToKM.R
\name{convertDegToKM}
\alias{convertDegToKM}
\title{Converts a set of coordinates in degrees into a new set of coordinates in The Universal Transverse Mercator (UTM) zone:37 coordinate system}
\usage{
convertDegToKM(
  loc,
  crs =
    sp::CRS("+units=km +proj=utm +zone=37 +ellps=clrk80 +towgs84=-160,-6,-302,0,0,0,0 +no_defs")
)
}
\arguments{
\item{loc}{A two column matrix of coordinates (The first column is longitude and the second column is latitude).}

\item{crs}{The coordinate reference system.}
}
\value{
A two column matrix of The Universal Transverse Mercator (UTM) zone:37 coordinate system (https://www.usgs.gov/faqs/what-does-term-utm-mean-utm-better-or-more-accurate-latitudelongitude).
}
\description{
Converts a set of coordinates in degrees into a new set of coordinates in The Universal Transverse Mercator (UTM) zone:37 coordinate system
}
\examples{
data("clusterData")
loc = cbind(clusterData$long, clusterData$lat)
locKM <- convertDegToKM(loc = loc)
head(locKM)
}
