% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridCountry.R
\name{gridCountry}
\alias{gridCountry}
\title{Creates a grid of locations within the national borders of a country of interest}
\usage{
gridCountry(admin0, res)
}
\arguments{
\item{admin0}{A SpatialPolygonsDataFrame object representing the national (admin0) level borders of the country}

\item{m}{A value representing the number of points that the longitude range of the country will be divided by}

\item{n}{A value representing the number of points that the latitude range of the country will be divided by}
}
\value{
A data frame containing the coordinates (in degrees and in kilometers) of a set of prediction points on a grid
}
\description{
Creates a grid of locations within the national borders of a country of interest
}
\examples{
\dontrun{
grid <- gridCountry(admin0 = admin0, m = m, n = n)
}
}
