% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRes.R
\name{predRes}
\alias{predRes}
\title{Predicts new outcomes for the response variable}
\usage{
predRes(obj, predCoords)
}
\arguments{
\item{obj}{The output object created by the "estimateModel" function that contains the TMB components for predictions}

\item{predCoords}{A matrix containing the coordinates of the prediction locations in kilometers}
}
\value{
A matrix called "PredictedResponses", containing the mean, median,
standard deviation and the lower and the upper bounds of 95% credible intervals
}
\description{
Predicts new outcomes for the response variable
}
\examples{
predRes(obj, predCoords)
}
