% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.res.R
\name{print.res}
\alias{print.res}
\title{Prints the output of estimateModel() function.}
\usage{
\method{print}{res}(est)
}
\arguments{
\item{est}{Distance matrix.}
}
\value{
Prints the model parameter estimates and the corresponding credible interval lengths.
}
\description{
Prints the output of estimateModel() function.
}
\examples{
\dontrun{
print.res(est = est)
}
}
